+++ NEWS.md
@@ -1,4 +1,6 @@
-# inSilecoRef 0.0.1.9000
+# inSilecoRef (devel)
+
+* Add new functions to quickly generate commonly used URL for R packages: `url_cran()`, `url_bioc()` and `url_gh()`.
 
 # inSilecoRef 0.0.1